requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
click==8.0.0
colorama==0.4.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
install==1.3.4
isort==5.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
